pyproject.toml
[project]
name = "dbbundle-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Gardner", email = "gardnmi@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.44.1",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "ruff>=0.11.2",
]

[project.optional-dependencies]
dev = [
  "databricks-connect==15.4.*",
  #"pyspark==3.5.*",
  #"delta-spark==3.2.*",
]

[project.scripts]
dbbundle-config = "dbbundle_config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
